# 激光塔和射击塔同时攻击多个敌人
- 创建管理被攻击对象的组件AttackHandleComponent
- 更改攻击逻辑，改为攻击多个敌人的逻辑
- Camera缩放
- 关掉AI的不必要碰撞
- 完善结束游戏逻辑

+++ Source/TowerDefense/Private/Component/ActorComp/TManStateAndBuffer.cpp
@@ -46,10 +46,10 @@ void UTManStateAndBuffer::TickComponent(float DeltaTime, ELevelTick TickType, FA
 	// ...
 }
 
-void UTManStateAndBuffer::ApplyHealthChange(AActor* Instigator,int Delta)
+void UTManStateAndBuffer::ApplyHealthChange(AActor* Instigator,float Delta)
 {
 	CurrentHealth += Delta;
-	CurrentHealth = FMath::Clamp(CurrentHealth, 0, ManState.MaxHealth);
+	CurrentHealth = FMath::Clamp(CurrentHealth, 0.0f, ManState.MaxHealth);
 	OnHealthChanged.Broadcast(Instigator,this, CurrentHealth,Delta);
 	//TODO: 添加死亡事件
 	if( CurrentHealth == 0)
